// src/store/types/rfp.actionTypes.ts
export enum RFP_ACTION_TYPES {
    
  // Generate RFP
  GENERATE_RFP_REQUEST = 'GENERATE_RFP_REQUEST',
  GENERATE_RFP_SUCCESS = 'GENERATE_RFP_SUCCESS',
  GENERATE_RFP_FAILURE = 'GENERATE_RFP_FAILURE',

  // Create RFP
  CREATE_RFP_REQUEST = 'CREATE_RFP_REQUEST',
  CREATE_RFP_SUCCESS = 'CREATE_RFP_SUCCESS',
  CREATE_RFP_FAILURE = 'CREATE_RFP_FAILURE',

  // Get All RFPs
  GET_ALL_RFPS_REQUEST = 'GET_ALL_RFPS_REQUEST',
  GET_ALL_RFPS_SUCCESS = 'GET_ALL_RFPS_SUCCESS',
  GET_ALL_RFPS_FAILURE = 'GET_ALL_RFPS_FAILURE',

  // Get RFP by ID
  GET_RFP_BY_ID_REQUEST = 'GET_RFP_BY_ID_REQUEST',
  GET_RFP_BY_ID_SUCCESS = 'GET_RFP_BY_ID_SUCCESS',
  GET_RFP_BY_ID_FAILURE = 'GET_RFP_BY_ID_FAILURE',

  // Update RFP
  UPDATE_RFP_REQUEST = 'UPDATE_RFP_REQUEST',
  UPDATE_RFP_SUCCESS = 'UPDATE_RFP_SUCCESS',
  UPDATE_RFP_FAILURE = 'UPDATE_RFP_FAILURE',

  // Delete RFP
  DELETE_RFP_REQUEST = 'DELETE_RFP_REQUEST',
  DELETE_RFP_SUCCESS = 'DELETE_RFP_SUCCESS',
  DELETE_RFP_FAILURE = 'DELETE_RFP_FAILURE',

  // Dashboard Stats
  GET_DASHBOARD_STATS_REQUEST = 'GET_DASHBOARD_STATS_REQUEST',
  GET_DASHBOARD_STATS_SUCCESS = 'GET_DASHBOARD_STATS_SUCCESS',
  GET_DASHBOARD_STATS_FAILURE = 'GET_DASHBOARD_STATS_FAILURE',

  // Clear Errors
  CLEAR_RFP_ERRORS = 'CLEAR_RFP_ERRORS',
}